chore: import MesheryPatternDeleteRequestBody struct from schema

+++ server/models/meshery_pattern_persister.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/meshery/meshkit/database"
 	"github.com/meshery/meshkit/models/patterns"
+	"github.com/meshery/schemas/models/v1beta1/pattern"
 )
 
 // MesheryPatternPersister is the persister for persisting
@@ -108,7 +109,6 @@ func (mpp *MesheryPatternPersister) GetMesheryCatalogPatterns(page, pageSize, se
 
 	var count int64
 	err = query.Model(&MesheryPattern{}).Count(&count).Error
-
 	if err != nil {
 		return nil, err
 	}
@@ -160,13 +160,15 @@ func (mpp *MesheryPatternPersister) DeleteMesheryPattern(id uuid.UUID) ([]byte, 
 }
 
 // DeleteMesheryPatterns takes in a meshery-patterns and delete those if exist
-func (mpp *MesheryPatternPersister) DeleteMesheryPatterns(patterns MesheryPatternDeleteRequestBody) ([]byte, error) {
+func (mpp *MesheryPatternPersister) DeleteMesheryPatterns(patterns pattern.MesheryPatternDeleteRequestBody) ([]byte, error) {
 	var deletedMaptterns []MesheryPattern
-	for _, pObj := range patterns.Patterns {
-		id := uuid.FromStringOrNil(pObj.ID)
-		pattern := MesheryPattern{ID: &id}
-		mpp.DB.Delete(&pattern)
-		deletedMaptterns = append(deletedMaptterns, pattern)
+	if patterns.Patterns != nil {
+		for _, pObj := range *patterns.Patterns {
+			id := pObj.Id
+			pattern := MesheryPattern{ID: &id}
+			mpp.DB.Delete(&pattern)
+			deletedMaptterns = append(deletedMaptterns, pattern)
+		}
 	}
 
 	return marshalMesheryPatterns(deletedMaptterns), nil
